* Halo1_CheApe: Made a util function for appending chars when a string doesn't end with that char (esp. for path seperators) * Halo1_CheApe: Addressed a sign mismatch in a comparison (int32 > size_t) * Halo1_CheApe: Changed SetQuadToDefaults to take the tag_index so it prints more meaningful debug info * Halo1: Uncommented out s_effect_postprocess_effect_fade_control and surrounded it in #if blocks so now it's just dead code (is this still a WIP?)

+++ OpenSauce/Halo1/Halo1_CheApe/TagGroups/yelo_scenario_definitions.cpp
@@ -35,7 +35,7 @@ namespace Yelo
 			// If the scenario is using upgraded script node sizes, clear it
 			// Users will need to recompile their scenario's scripts
 			tag_data& hs_syntax_data = scnr->hs_syntax_data;
-			if(hs_syntax_data.size > Scripting::GetTotalScenarioHsSyntaxData())
+			if((size_t)hs_syntax_data.size > Scripting::GetTotalScenarioHsSyntaxData())
 				tag_data_delete(hs_syntax_data); // TODO: does hs_syntax_data need to be initialized to GetTotalScenarioHsSyntaxData instead of just zero?
 		}
